Revert "Working with API First Submission"
