+++ blocks/modal.css
@@ -106,54 +106,6 @@
   pointer-events: none;
 }
 
-.modal__delete-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 630px;
-  height: 172px;
-}
-.modal__delete-close-button {
-  background-image: url("../icon/close-icon-light.svg");
-  right: -42px;
-  top: -5px;
-  opacity: 1;
-}
-.modal__delete-button {
-  background-color: rgba(255, 101, 101, 1);
-  padding: 18px 32px;
-  gap: 8px;
-  border-radius: 8px;
-  width: 156px;
-  height: 52px;
-  border: none;
-  margin-right: 20px;
-}
-.modal__delete-button:hover {
-  cursor: url("../icon/pointer.svg"), pointer;
-  background-color: rgba(255, 51, 51, 1);
-}
-.modal__cancel-button {
-  border-color: rgba(33, 33, 33, 1);
-  padding: 18px 32px;
-  gap: 8px;
-  border-radius: 8px;
-  width: 156px;
-  height: 52px;
-  border: solid 1px;
-}
-.modal__cancel-button:hover {
-  cursor: url("../icon/pointer.svg"), pointer;
-  background-color: rgba(33, 33, 33, 1);
-  color: white;
-}
-
-.modal__delete-title {
-  font-size: 24px;
-  font-weight: 500;
-  line-height: 100%;
-}
-
 .modal__preview-content {
   padding: 0;
   background-color: transparent;
@@ -234,25 +186,5 @@
   .modal__close-button_preview {
     top: -40px;
     right: 0;
-  }
-
-  .modal__delete-container {
-    width: 300px;
-    height: fit-content;
-    text-align: center;
-  }
-  .modal__delete-title {
-    margin: 10px 0 12px;
-    font-size: 20px;
-    line-height: 120%;
-  }
-  .modal__delete-form {
-    gap: 12px;
-    display: flex;
-    flex-direction: column;
-  }
-  .modal__delete-close-button {
-    right: -5px;
-    top: -55px;
   }
 }
