+++ blocks/profile.css
@@ -11,36 +11,6 @@
   border-radius: 8px;
 }
 
-.profile__avatar-container {
-  position: relative;
-  margin-right: 20px;
-  height: 190px;
-  width: 190px;
-}
-
-.profile__avatar-button {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: flex-end;
-  justify-content: center;
-  visibility: hidden;
-  background-color: rgba(0, 0, 0, 0.5);
-  color: white;
-  border-radius: 8px;
-  border: none;
-}
-
-.profile__avatar-container:hover .profile__avatar-button {
-  visibility: visible;
-  cursor: url("../icon/pointer.svg"), pointer;
-}
-
-.profile__avatar-button::after {
-  content: "Edit Avatar";
-}
-
 .profile__column {
   display: flex;
   flex-direction: column;
@@ -139,21 +109,6 @@
     margin: 0 0 12px;
     height: 80px;
     width: 80px;
-  }
-  .profile__avatar-button::after {
-    content: "";
-  }
-
-  .profile__avatar-button {
-    visibility: visible;
-    background-color: (33, 33, 33, 0.6);
-    top: 52px;
-    left: 105px;
-    border-radius: 6px;
-    padding: 4px;
-    position: absolute;
-    width: auto;
-    height: auto;
   }
   .profile__new-post {
     margin: 0;
