+++ index.html
@@ -10,33 +10,25 @@
     />
     <meta name="keywords" content="photography portfolio spots photo" />
     <link rel="shortcut icon" href="./favicon.ico" type="image/x-icon" />
+    <link rel="stylesheet" href="pages/index.css" />
     <title>Spots</title>
   </head>
   <body>
     <div class="page">
       <header class="header">
         <img
-          src="<%= require('./icon/spots-logo.svg')%>"
+          src="./icon/spots-logo.svg"
           alt="Spots Logo"
           class="header__logo"
         />
       </header>
       <main class="content">
         <section class="profile page__section">
-          <div class="profile__avatar-container">
-            <button class="profile__avatar-button">
-              <img
-                src="<%= require('./icon/pencil-light.svg')%>"
-                alt="Profile Pencil Icon"
-                class="profile__pencil-light"
-              />
-            </button>
-            <img
-              src="<%= require('./images/avatar.jpg')%>"
-              alt="Profile Avatar"
-              class="profile__avatar"
-            />
-          </div>
+          <img
+            src="./images/avatar.jpg"
+            alt="Profile Avatar"
+            class="profile__avatar"
+          />
           <div class="profile__column">
             <h1 class="profile__name">Bessi Coleman</h1>
             <p class="profile__description">Civil Aviator</p>
@@ -44,7 +36,7 @@
           </div>
           <button type="button" class="profile__new-post">
             <img
-              src="<%= require('./icon/plus-icon.svg')%>"
+              src="./icon/plus-icon.svg"
               alt="plus icon"
               class="profile__new-post-icon"
             />New Post
@@ -142,41 +134,6 @@
           </form>
         </div>
       </div>
-      <div class="modal" id="avatar-modal">
-        <div class="modal__container">
-          <button type="button" class="modal__close-button"></button>
-          <h2 class="modal__title">Edit profile avatar</h2>
-          <form name="new-avatar-form" class="modal__form">
-            <label for="new-avatar-link-input" class="modal__label"
-              >Profile Picture
-              <input
-                id="avatar-link-input"
-                type="url"
-                class="modal__input"
-                placeholder="Paste the link to an image"
-                name="link"
-                required />
-              <span id="avatar-link-input-error" class="modal__error"></span
-            ></label>
-            <button type="submit" class="modal__save-button">Save</button>
-          </form>
-        </div>
-      </div>
-      <div class="modal" id="delete-modal">
-        <div class="modal__container modal__delete-container">
-          <button
-            type="button"
-            class="modal__close-button modal__delete-close-button"
-          ></button>
-          <h2 class="modal__title modal__delete-title">
-            Are you sure you want to delete this image?
-          </h2>
-          <form class="modal__form modal__delete-form" id="delete-form">
-            <button type="submit" class="modal__delete-button">Delete</button>
-            <button type="button" class="modal__cancel-button">Cancel</button>
-          </form>
-        </div>
-      </div>
     </div>
     <div class="modal" id="preview-modal">
       <div class="modal__preview-content">
@@ -188,5 +145,7 @@
         <p class="modal__caption"></p>
       </div>
     </div>
+    <script defer src="./scripts/index.js"></script>
+    <script defer src="./scripts/validation.js"></script>
   </body>
 </html>
