--- package.json
@@ -1,38 +1,0 @@
-{
-  "name": "se_project_spots",
-  "version": "1.0.0",
-  "description": "Spots Project",
-  "main": "index.js",
-  "scripts": {
-    "build": "webpack --mode production",
-    "dev": "webpack serve --mode development"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/K-Nguyen-92/se_project_spots.git"
-  },
-  "author": "Khoa Nguyen",
-  "license": "ISC",
-  "devDependencies": {
-    "@babel/core": "7.18.9",
-    "@babel/preset-env": "7.18.9",
-    "autoprefixer": "10.4.7",
-    "babel-loader": "8.2.5",
-    "clean-webpack-plugin": "4.0.0",
-    "css-loader": "6.7.1",
-    "cssnano": "5.1.12",
-    "html-webpack-plugin": "5.5.0",
-    "mini-css-extract-plugin": "2.6.1",
-    "postcss-loader": "7.0.1",
-    "webpack": "5.76.0",
-    "webpack-cli": "4.10.0",
-    "webpack-dev-server": "4.9.3"
-  },
-  "dependencies": {
-    "core-js": "3.23.5"
-  },
-  "bugs": {
-    "url": "https://github.com/K-Nguyen-92/se_project_spots/issues"
-  },
-  "homepage": "https://github.com/K-Nguyen-92/se_project_spots#readme"
-}
