--- src/utils/helpers.js
@@ -1,12 +1,0 @@
-export function setButtonText(
-  button,
-  isLoading,
-  defaultText = "Save",
-  loadingText = "Saving..."
-) {
-  if (isLoading) {
-    button.textContent = loadingText;
-  } else {
-    button.textContent = defaultText;
-  }
-}
